+++ notificare-push/src/main/java/re/notifica/push/internal/ServiceManager.kt
@@ -2,14 +2,14 @@ package re.notifica.push.internal
 
 import re.notifica.InternalNotificareApi
 import re.notifica.internal.AbstractServiceManager
-import re.notifica.models.NotificareTransport
+import re.notifica.push.models.NotificareTransport
 
 @InternalNotificareApi
 public abstract class ServiceManager : AbstractServiceManager() {
 
     public abstract val transport: NotificareTransport
 
-    public abstract fun requestPushToken()
+    public abstract suspend fun getPushToken(): String
 
     internal companion object {
         private const val GMS_FQN = "re.notifica.push.gms.internal.ServiceManager"