+++ notificare-push/src/main/java/re/notifica/push/models/NotificareTransport.kt
@@ -1,0 +1,19 @@
+package re.notifica.push.models
+
+import android.os.Parcelable
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+@JsonClass(generateAdapter = false)
+public enum class NotificareTransport : Parcelable {
+    @Json(name = "Notificare")
+    NOTIFICARE,
+
+    @Json(name = "GCM")
+    GCM,
+
+    @Json(name = "HMS")
+    HMS
+}