+++ notificare/src/main/java/re/notifica/models/NotificareDevice.kt
@@ -2,7 +2,6 @@ package re.notifica.models
 
 import android.os.Parcelable
 import com.squareup.moshi.JsonClass
-import java.util.Date
 import kotlinx.parcelize.Parcelize
 import org.json.JSONObject
 import re.notifica.Notificare
@@ -17,16 +16,8 @@ public data class NotificareDevice internal constructor(
     val userId: String?,
     val userName: String?,
     val timeZoneOffset: Double,
-    val osVersion: String,
-    val sdkVersion: String,
-    val appVersion: String,
-    val deviceString: String,
-    val language: String,
-    val region: String,
-    val transport: NotificareTransport,
     val dnd: NotificareDoNotDisturb?,
     val userData: NotificareUserData,
-    val lastRegistered: Date,
 ) : Parcelable {
 
     public fun toJson(): JSONObject {