+++ sample/src/main/java/re/notifica/sample/MainApplication.kt
@@ -35,6 +35,17 @@ class MainApplication : Application(), Notificare.Listener {
         Notificare.addListener(this)
 
         Notificare.launch()
+
+        applicationScope.launch {
+            try {
+                if (Notificare.canEvaluateDeferredLink()) {
+                    val evaluated = Notificare.evaluateDeferredLink()
+                    Timber.i("deferred link evaluation = $evaluated")
+                }
+            } catch (e: Exception) {
+                Timber.e(e, "Failed to evaluate the deferred link.")
+            }
+        }
     }
 
     override fun onReady(application: NotificareApplication) {
@@ -68,11 +79,9 @@ class MainApplication : Application(), Notificare.Listener {
             val userName = getString(R.string.sample_user_name).ifBlank { null }
 
             try {
-                Timber.i("Registering device")
-                Notificare.device().register(userId, userName)
-                Timber.i("Registered device successfully.")
+                Notificare.device().updateUser(userId, userName)
             } catch (e: Exception) {
-                Timber.e(e, "Failed to register device.")
+                Timber.e(e, "Failed to update the user.")
             }
         }
     }