+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # CHANGELOG
 
+## Upcoming release
+
+- Fix padding discrepancies in alert dialog with actions
+- Device identifiers become long-lived
+
+#### Breaking changes
+
+- `NotificareDevice.id` attribute no longer contains the push token. Use `Notificare.push().subscriptionId` instead.
+- The `NotificareDevice` data model was reduced to only publicly relevant attributes.
+- `onDeviceRegistered` only triggers once, when the device is created.
+- `NotificareTransport` was moved to the push module.
+- Drops support for the monetize module.
+- Drops support for Huawei Mobile Services.
+- Drops support for v1 passes in-app wallet.
+
+## 3.9.0
+
+- Add support for deferred links
+
 ## 3.8.0
 
 - Prevent processing location updates too close to the last known location